# docs/openapi/doco-openapi-v1.yaml
openapi: 3.1.0
info:
  title: Doco Open API
  version: 1.0.0
  description: Doco 知识库、文档正文、块和附件开放 API。标准正文格式为 Tiptap/ProseMirror JSON。
servers:
  - url: https://example.com/api/v1
security:
  - bearerAuth: []
...
paths:
...
  /knowledge-bases/{id}/export:
    get:
      tags: [KnowledgeBases]
      operationId: exportKnowledgeBase
      parameters: [{ $ref: '#/components/parameters/ResourceId' }]
      responses: { '200': { description: ZIP 文件, content: { application/zip: { schema: { type: string, contentEncoding: binary } } } }, '404': { $ref: '#/components/responses/Error' } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: 'doco_<token_id>_<secret>' }
  parameters:
    ResourceId: { in: path, name: id, required: true, schema: { type: string } }
...
  headers:
    RequestId: { schema: { type: string }, description: 服务端或调用方提供的请求 ID }
...
  responses:
...
    Error:
      description: 错误
      headers: { X-Request-Id: { $ref: '#/components/headers/RequestId' } }
      content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } }
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error, request_id]
      properties:
        error: { type: object, required: [type, code, message], properties: { type: { type: string }, code: { type: string }, message: { type: string }, details: { type: object } } }
        request_id: { type: string }
...